Rename and move event
